requirements.txt
flask~=1.1.2
python-dotenv~=0.15.0
psycopg2~=2.8.6
flask-login~=0.5.0
flask-sqlalchemy~=2.4.4
flask-session~=0.3.2
bcrypt~=3.2.0
pyyaml~=5.4.1
werkzeug~=1.0.1

pytest~=6.2.2
flexmock~=0.10.4
pep8~=1.7.1
